requirements.txt
Sphinx==8.0.*
myst-parser==4.0.*
furo==2024.8.*
sphinx-book-theme==1.1.*
sphinx-autoapi==3.2.*
sphinx_collapse==0.1.*
sphinx-copybutton==0.5.*
numpy==1.26.*
scipy==1.13.*
matplotlib==3.9.*
seaborn==0.13.*
colorednoise==2.2.*
pandas==2.2.*
tqdm==4.66.*
PyYAML==6.0.1
bottleneck==1.3.6
alphawaves==0.5
mne==1.8.*